+++ app/Modules/Config/Resources/Views/admin_index.blade.php
@@ -15,6 +15,7 @@
     <ul class="nav nav-tabs" role="tablist">
         <li role="presentation" class="active"><a href="#tab_general" aria-controls="tab_general" role="tab" data-toggle="tab">{{ trans('app.general') }}</a></li>
         <li role="presentation"><a href="#tab_services" aria-controls="tab_services" role="tab" data-toggle="tab">{{ trans('app.services') }}</a></li>
+        <li role="presentation"><a href="#tab_meta" aria-controls="tab_meta" role="tab" data-toggle="tab">{{ trans('app.metainfo') }}</a></li>
         <li role="presentation"><a href="#tab_theme" aria-controls="tab_theme" role="tab" data-toggle="tab">{{ trans('app.theme') }}</a></li>
     </ul>
 
@@ -53,6 +54,20 @@
             <hr>
 
             {!! Form::smartTextarea('app::analytics', trans('config::analytics'), false) !!}
+        </div>
+		<div role="tabpanel" class="tab-pane" id="tab_meta">
+            {!! Form::smartText('app::author', trans('config::meta_author'), false) !!}
+            <p class="help-block  text-right">{!! trans('config::meta_author_info') !!}</p>
+
+            <hr>
+
+            {!! Form::smartText('app::keywords', trans('config::meta_keywords'), false) !!}
+            <p class="help-block  text-right">{!! trans('config::meta_keywords_info') !!}</p>
+
+            <hr>
+
+            {!! Form::smartTextarea('app::description', trans('config::meta_description'), false) !!}
+            <p class="help-block  text-right">{!! trans('config::meta_description_info') !!}</p>
         </div>
         <div role="tabpanel" class="tab-pane" id="tab_theme">
             {!! Form::smartGroupOpen('app::theme', trans('app.theme')) !!}