+++ app/Modules/MorpheusTheme/Resources/Views/layout.blade.php
@@ -6,11 +6,14 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <meta name="generator" content="Contentify">
-    <meta name="base-url" content="{!! url('/') !!}">
-    <meta name="asset-url" content="{!! asset('') !!}">
-    <meta name="csrf-token" content="{!! Session::get('_token') !!}">
-    <meta name="locale" content="{!! Config::get('app.locale') !!}">
-    <meta name="date-format" content="{!! trans('app.date_format') !!}">
+    <meta name="base-url" content="{{ url('/') }}">
+    <meta name="asset-url" content="{{ asset('') }}">
+    <meta name="csrf-token" content="{{ Session::get('_token') }}">
+    <meta name="locale" content="{{ Config::get('app.locale') }}">
+    <meta name="date-format" content="{{ trans('app.date_format') }}">
+    <meta name="author" content="{{ Config::get('app.author') }}">
+    <meta name="keywords" content="{{ Config::get('app.keywords') }}">
+    <meta name="description" content="{{ Config::get('app.description') }}">
     {!! HTML::metaTags($metaTags) !!}
     @if ($openGraph)
         {!! HTML::openGraphTags($openGraph) !!}