+++ app/Modules/Pages/Http/Controllers/FragmentWidget.php
@@ -3,7 +3,7 @@
 namespace App\Modules\Pages\Http\Controllers;
 
 use App\Modules\Pages\Fragment;
-use Exception;
+use InvalidArgumentException;
 use Widget;
 
 class FragmentWidget extends Widget
@@ -14,7 +14,7 @@ class FragmentWidget extends Widget
         if (isset($parameters['id'])) {
             $id = $parameters['id'];
         } else {
-            throw new Exception('Fragment Widget: ID parameter missing.');
+            throw new InvalidArgumentException('Fragment Widget: ID parameter missing.');
         }
 
         $fragment = Fragment::findOrFail($id);