+++ contentify/Traits/SlugTrait.php
@@ -6,7 +6,7 @@ use DB;
 use Exception;
 use Str;
 
-/*
+/**
  * Use this trait to "extend" Eloquent models so they can create "smart" slugs.
  * Classes that include this trait should (must) inherit from the \Contentify\Models\BaseModel class.
  */