+++ resources/lang/es/app.php
@@ -202,6 +202,7 @@ return [
     'nothing_here'      => 'Lo siento, aún no hay contenido aquí. Por favor, inténtelo de nuevo más tarde.',
     'back'              => 'Atrás',
     'general'           => 'General',
+    'metainfo'			=> 'Meta Información',
     'services'          => 'Servicios',
     'theme'             => 'Tema',
     'theme_christmas'   => 'Modo navidad (Nieve)',