+++ resources/lang/fr/app.php
@@ -202,6 +202,7 @@ return [
     'nothing_here'      => 'Désoler, il n\'y a pas encore de contenu. Veuillez réessayer plus tard.',
     'back'              => 'Retour',
     'general'           => 'Généralement',
+	'metainfo'			=> 'Méta-Information',
     'services'          => 'Services',
     'theme'             => 'Thème',
     'theme_christmas'   => 'Mode Nöel (Flocon)',