+++ changelog.md
@@ -1,3 +1,7 @@
 ## Changelog - v2.6
 
+- Added meta information (author, description and keywords) to the website settings (thanks to kamohy!)
+- Added polls module  which also includes a widget. Take a look at the `README.md` in the modules folder to learn more.
+- Display info text if a website is empty, because it does not have any content yet
+- Fixed bugs 
 - *To be announced*